♻️  refactor(routes): reorganize routes, change root to auth and app to root. reorganize folder structure
